+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+#### Public API Updates!
+- Add `Say.progress_line` and `Say#progress_line` for printing a message along with a given `index` indicator. `Say.progress` now uses this internally so that calling `say` on an Interval object automatically includes the `index` indicator.
+
 ## [0.3.0] - 2023-06-02
 
 ### BREAKING CHANGES: